Add files via upload
This Project is a simple project for my CS 230 class. The task at hand was to create an instance of a clock; however, it does not move in real-time. The clock can be set to whatever time you wish it to be, and it operates in 12-hour and 24-hour formats. It also tracks AM and PM times for the 12-hour time zones.  This project was for a mock company that wished to display its time to users not only in 12-hour time zones but in 24-hour time zones as well.   This is important for companies who wish to have a universal audience or people in other countries who operate with 24-hour time zones who need to operate within the companies time zones as well.